# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyan Travels API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WaitlistEntry {
  id: number;
  email: string;
  /** @nullable */
  createdAt?: string | null;
}

export interface WaitlistInput {
  email: string;
  name?: string;
}

export interface WaitlistCount {
  count: number;
}

export interface ErrorResponse {
  error: string;
}