# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gyan Travels API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: waitlist
    description: Waitlist operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /waitlist:
    post:
      operationId: joinWaitlist
      tags: [waitlist]
      summary: Join the waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "201":
          description: Successfully joined waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
        "400":
          description: Invalid email or already on waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /waitlist/count:
    get:
      operationId: getWaitlistCount
      tags: [waitlist]
      summary: Get total waitlist count
      responses:
        "200":
          description: Waitlist count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistCount"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: ["string", "null"]
      required: [id, email]
    WaitlistInput:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
      required: [email]
    WaitlistCount:
      type: object
      properties:
        count:
          type: integer
      required: [count]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]